# Documentation/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Birøkt API
  description: |
    RESTful API for birøkt-systemet (beekeeping management system).
    
    ## Authentication
    All endpoints (except auth endpoints) require a Bearer token in the Authorization header.
    
    ## Rate Limiting
    - Auth endpoints: 5 requests/minute per IP
    - Read endpoints: 100 requests/minute per user
    - Write endpoints: 30 requests/minute per user
    
  version: 1.0.0
  contact:
    name: Birøkt Support
    email: support@birokt.no
  license:
    name: Proprietary
servers:
  - url: https://api.birokt.no/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
...
    Hive:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hive_number:
          type: string
          example: K12
        qr_code:
          type: string
          nullable: true
        apiary:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        status:
          type: string
          enum: [active, nuc, inactive, dead, sold]
        strength:
          type: string
          enum: [weak, medium, strong]
          nullable: true
        hive_type:
          type: string
          enum: [langstroth, topbar, warre]
        box_count:
          type: integer
        queen:
          type: object
          nullable: true
          properties:
            year:
              type: integer
            marked:
              type: boolean
            color:
              type: string
              enum: [white, yellow, red, green, blue]
            race:
              type: string
        current_frames:
          type: object
          properties:
            brood:
              type: integer
            honey:
              type: integer
        last_inspection:
          type: object
          nullable: true
          properties:
            date:
              type: string
              format: date-time
            health_status:
              type: string
            strength:
              type: string
        stats:
          type: object
          properties:
            total_inspections:
              type: integer
            total_production_kg:
              type: number
              format: float
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /hives:
    get:
      tags:
        - Hives
      summary: Get all hives for user
      security:
        - bearerAuth: []
      parameters:
        - name: apiary_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, nuc, inactive, dead, sold]
        - name: strength
          in: query
          schema:
            type: string
            enum: [weak, medium, strong]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of hives
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Hive'
                  meta:
                    type: object
                    properties:
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Hives
      summary: Create new hive
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - apiary_id
                - hive_number
              properties:
                apiary_id:
                  type: string
                  format: uuid
                hive_number:
                  type: string
                hive_type:
                  type: string
                  enum: [langstroth, topbar, warre]
                status:
                  type: string
                  enum: [active, nuc, inactive]
                queen:
                  type: object
                  properties:
                    year:
                      type: integer
                    marked:
                      type: boolean
                    color:
                      type: string
                    race:
                      type: string
      responses:
        '201':
          description: Hive created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Hive'
  /hives/{id}:
    get:
      tags:
        - Hives
      summary: Get hive by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Hive details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Hive'
...